requirements.txt
nltk==3.4.5
Pillow==6.1.0
pytesseract==0.3.0
six==1.12.0
